[refactor] 프로필 이미지 등록 라우트 및 페이지 제거 #43
