+++ src/components/influencerMypage/profile/ProfileFanCard.vue
@@ -16,7 +16,7 @@ const goToEditFanCard = () => {
   <div class="px-5 mt-6 mb-6 space-y-2">
     <!-- 타이틀 + 아이콘 -->
     <div class="flex items-center justify-between cursor-pointer" @click="goToEditFanCard">
-      <span class="text-sm font-medium text-base">팬카드 이미지</span>
+      <span class="text-sm text-base">팬카드 이미지</span>
       <img :src="Icon" alt=">" class="w-3 h-3 rotate-180" />
     </div>
