+++ src/components/influencerMypage/profile/ProfileImage.vue
@@ -1,16 +1,22 @@
 <script setup>
-import { useRouter } from 'vue-router'
-
+import { ref } from 'vue'
 import Ellipse from '@/assets/InfluencerProfile/Ellipse.svg'
 import PlusIcon from '@/assets/InfluencerProfile/Plus.svg'
-
 import { useInfluencerStore } from '@/stores/influencerStore'
+
 const influencerStore = useInfluencerStore()
+const fileInputRef = ref(null)
 
-const router = useRouter()
+const triggerUpload = () => {
+  fileInputRef.value?.click()
+}
 
-const goToEditImage = () => {
-  router.push('/influencer/profile/edit-image')
+const onFileChange = (e) => {
+  const file = e.target.files[0]
+  if (file) {
+    const preview = URL.createObjectURL(file)
+    influencerStore.setProfileImage(preview)
+  }
 }
 </script>
 
@@ -19,14 +25,23 @@ const goToEditImage = () => {
     <div class="relative w-24 h-24">
       <!-- 프로필 이미지 -->
       <img
-        :src="influencerStore.profileImage || defaultImage"
+        :src="influencerStore.profileImage"
         alt="프로필 이미지"
         class="w-24 h-24 rounded-full object-cover border"
       />
-      <div class="absolute bottom-0 right-0 w-6 h-6 cursor-pointer" @click="goToEditImage">
+      <div class="absolute bottom-0 right-0 w-6 h-6 cursor-pointer" @click="triggerUpload">
         <img :src="Ellipse" alt="버튼 배경" class="absolute w-full h-full" />
         <img :src="PlusIcon" alt="플러스" class="absolute w-3 h-3 top-1.5 left-1.5" />
       </div>
+
+      <!-- 숨겨진 파일 업로드 input -->
+      <input
+        type="file"
+        accept="image/*"
+        ref="fileInputRef"
+        class="hidden"
+        @change="onFileChange"
+      />
     </div>
   </div>
 </template>
