+++ src/router/influencerMypage.js
@@ -3,7 +3,6 @@ import ProfilePage from '@/pages/influencerMypage/ProfilePage.vue'
 import EditNamePage from '@/pages/influencerMypage/EditNamePage.vue'
 import EditCategoryPage from '@/pages/influencerMypage/EditCategoryPage.vue'
 import EditIntroPage from '@/pages/influencerMypage/EditIntroPage.vue'
-import EditProfileImagePage from '@/pages/influencerMypage/EditProfileImagePage.vue'
 import EditFanCardPage from '@/pages/influencerMypage/EditFanCardPage.vue'
 
 export default [
@@ -31,11 +30,6 @@ export default [
     path: '/influencer/profile/edit-intro',
     name: 'EditIntro',
     component: EditIntroPage,
-  },
-  {
-    path: '/influencer/profile/edit-image',
-    name: 'EditProfile',
-    component: EditProfileImagePage,
   },
   {
     path: '/influencer/profile/edit-fancard',
